Fix Home Page Rendering Issues
each discovery component was responsible for fetching the data, which caused a tremendous amount of rerending. now the home page is deciding what users to render and then passing everything along at once to the discovery components. updated some css and the discovery tabs are a little more selective in content