Improve Birthday Form
Who knew this would take so long... been working on this since Sunday. However: now I've added styling, removed a bug that would throw the under 18 error (which should be impossible), and now the user can go back and form from the next form and update the birthday. Still have a bug where the user's birthday is turning into the day before entered...