Refactor Messages
I've refactored enough so that when you send a new message from the messages page, it appears. I'd like to refactor more since right now I'm passing all messages as props to each thread, and then only filtering the messages down within the thread. a better solution would be to filter down the messages from the highest component and only pass to each messages thread what is pertinent. This will also allow me to sort by which user has been messaged with most recently. Since I want to focus on refactoring the home page (which will take a while), I'm going to move on from this for now