Write Nearby Users In Backend
This is imperfect, but I wanted a solution to pair down the pool of users made available to the home page. Since the users database is indexed on location, this is a quick operation. Now I can perform the more expensive Google Distance Matrix API request on the front end if I need to