Refactor Home Page Fetching
It is now using the find local users function. It's clear that I need to add the MatchPercentages table since the loading time is still a little slow. This table will store match percentages and only recalculate if one of the users has responded to a question within the last day.