Write Random Answered Question AJAX
also quickly created a form and view component. we don't want to render a form version of the question unless the user is the current user. we'll render a view of the random answer to a question on other profiles and spawn a random question to answer for the current user