Fixed User Creation Bug
The bug was: if you created an account, logged out, and went to create a new account right away, the selections from the previous gender and ethnicity forms were preserved. With a high amount of debuggers, I determined I was actually unintentionally modifying the default _newUser constant in my reducer. This is because the components were written pretty early on in the process, and I was a little sloppy with my props / state understanding management. The quick fix is that I'm returning a new blank slice of state when the user logs in / out. I also uncovered a bug where the genders and ethnicities weren't validating properly because of set / array Ruby confusion